rename attractor files
