+++ shaders/attractor-sprite.wgsl
